change field data type of 'totalCount' from Long to Int

+++ packages/server-scala/src/main/scala/modules/pagination/PaginatedResult.scala
@@ -1,3 +1,3 @@
 package modules.pagination
 
-case class PaginatedResult[T](totalCount: Long, entities: List[T], hasNextPage: Boolean)
+case class PaginatedResult[T](totalCount: Int, entities: List[T], hasNextPage: Boolean)
